// src/services/api/generated/models/role-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `SUPER_ADMIN` - Super Administrator * `ADMIN` - Administrator * `NORMAL_USER` - Normal User
 */

export const RoleEnum = {
    SUPER_ADMIN: 'SUPER_ADMIN',
    ADMIN: 'ADMIN',
    NORMAL_USER: 'NORMAL_USER'
} as const;

export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];