// src/services/api/generated/models/patched-user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { RoleEnum } from './role-enum';

/**
 * Basic user serializer.
 */
export interface PatchedUser {
    'id'?: number;
    'email'?: string;
    /**
     * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
     */
    'username'?: string;
    'role'?: RoleEnum;
    'is_active'?: boolean;
    'last_login'?: string | null;
}